% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinKF.R
\name{LinKF}
\alias{LinKF}
\title{Linear Kalman filtering}
\usage{
LinKF(mZ, mX, Gam, Bet, mG, h0, P0, mX0)
}
\arguments{
\item{mZ}{a T by p matrix representing observations}

\item{mX}{a T by u matrix representing explanatory variables}

\item{Gam}{a p+q by q matrix representing the coefficients $(H' F')'$ of the latent states}

\item{Bet}{a p+q by u matrix representing the coefficients $(A' B')'$ of $X$}

\item{mG}{a p+q by p+q matrix representing the coefficients of errors $(R' Q')'$}

\item{h0}{a q by 1 matrix representing the initial state}

\item{P0}{a q by q matrix representing the initial state's covariance matrix}

\item{mX0}{a u by 1  matrix representing $x_0$ at time 0}
}
\value{
A list consisting of the following components:

\item{h1}{T by q the estimated conditional expectations of the latent states given t-1 (a priori)}
\item{h2}{T by q the estimated conditional expectations of the latent states given t (a posteriori)}
\item{P1}{T by q by q array of covariances of the a priori states}
\item{P2}{T by q by q array of covariances of the a posteriori states}
\item{v}{T by p residuals in the measurement eq.}
\item{S}{T by p by p array of covariances of z}
\item{K}{T by q by p array of Kalman gains}
}
\description{
\code{LinKF} implements the the linear Kalman filtering algorithm.
}
\details{
The function takes several arguments that specify the linear state space model as well as the data
and returns a list consisting of the estimates from the Kalman filtering algorithm.

The corresponding linear state space model is as follows:

\deqn{(z_t', h_{t+1}') = Gam h_t + Bet x_t + G e_t,}
where \eqn{Gam = (H', F')'}, \eqn{Bet = (A', B')'}, and \eqn{G = (R', Q')'}
}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
