% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimKF.R
\name{SimKF}
\alias{SimKF}
\title{Simulate from a linear state space model.}
\usage{
SimKF(mX, Gam, Bet, mG, h0, mX0)
}
\arguments{
\item{mX}{a T by u matrix representing the explanatory variable $X$}

\item{Gam}{a p+q by q matrix representing the coefficients $(H' F')'$ of the latent states}

\item{Bet}{a p+q by u matrix representing the coefficients $(A' B')'$ of $X$}

\item{mG}{a p+q by p+q matrix representing the coefficients of errors $(R' Q')'$}

\item{h0}{a q by 1 matrix representing the initial state}

\item{mX0}{a u by 1  matrix representing $x_0$ at time 0}
}
\value{
A list consisting of the following components:

\item{Z}{T by p  the observations}
\item{h}{T by q the latent process of the states}
}
\description{
\code{SimKF} implements the simulation from a linear state space model.
}
\details{
The function takes several arguments that specify the linear state space model
and returns a list consisting of the simulated latent process and the observations.

The corresponding linear state space model is as follows:

\deqn{(z_t', h_{t+1}') = Gam h_t + Bet x_t + G e_t,}
where \eqn{Gam = (H', F')'}, \eqn{Bet = (A', B')'}, and \eqn{G = (R', Q')'}

The seed can be set before running the function.
}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
